Master optimization

+++ api/db/order.go
@@ -13,19 +13,19 @@ import (
 
 var orderCollection = Db.Collection("Order")
 
-func getOneOrder(id primitive.ObjectID) models.Order {
+func getOneOrder(id primitive.ObjectID) (models.Order, error) {
 	data := orderCollection.FindOne(context.Background(), bson.M{"_id": id})
 
 	var order models.Order
 	err := data.Decode(&order)
 	if err != nil {
-		fmt.Println("Unable to find Order with id", id, ".Error: ", err)
+		return models.Order{}, err
 	}
 
-	return order
+	return order, nil
 }
 
-func getAllOrders() []models.Order {
+func getAllOrders() ([]models.Order, error) {
 	cursor, err := orderCollection.Find(context.Background(), bson.M{})
 	if err != nil {
 		fmt.Println("Unable to find all orders. Error: ", err)
@@ -41,22 +41,23 @@ func getAllOrders() []models.Order {
 		orders = append(orders, order)
 	}
 
-	return orders
+	if len(orders) == 0 {
+		return []models.Order{}, errors.New("no order data is present in ddatabase")
+	}
+
+	return orders, nil
 }
 
 func insertOneOrder(order models.Order) (*mongo.InsertOneResult, error) {
 	inserted, err := orderCollection.InsertOne(context.Background(), order)
 	if err != nil {
-		fmt.Println("Unable to insert one orders. Error: ", err)
 		return &mongo.InsertOneResult{}, err
 	}
 
-	fmt.Println("Order created with id : ", inserted.InsertedID)
-	fmt.Println("Created Order : ", order)
 	return inserted, nil
 }
 
-func updateOneOrder(id primitive.ObjectID, updatedValues primitive.M) models.Order {
+func updateOneOrder(id primitive.ObjectID, updatedValues primitive.M) (models.Order, error) {
 
 	filter := bson.M{"_id": id}
 
@@ -64,54 +65,61 @@ func updateOneOrder(id primitive.ObjectID, updatedValues primitive.M) models.Ord
 
 	updated, err := orderCollection.UpdateOne(context.Background(), filter, update)
 	if err != nil {
-		fmt.Println("Unable to update one order. Error: ", err)
+		return models.Order{}, err
 	}
 
 	if updated.ModifiedCount > 0 {
-		fmt.Println("Order successfully updated with id: ", id)
+		updatedOrder, _ := getOneOrder(id)
+		return updatedOrder, nil
 	}
-
-	updatedOrder := getOneOrder(id)
-
-	return updatedOrder
+	return models.Order{}, errors.New("no matching order found")
 }
 
-func deleteOneOrder(id primitive.ObjectID) models.Order {
+func deleteOneOrder(id primitive.ObjectID) (models.Order, error) {
 
-	deletedOrder := getOneOrder(id)
+	deletedOrder, _ := getOneOrder(id)
+	if deletedOrder == (models.Order{}) {
+		return models.Order{}, errors.New("no matching order found")
+	}
+
 	filter := bson.M{"_id": id}
 
 	deleted, err := orderCollection.DeleteOne(context.Background(), filter)
 	if err != nil {
-		fmt.Println("Unable to delete one order. Error: ", err)
+		return models.Order{}, errors.New("Unable to delete one order. Error : " + err.Error())
 	}
 
-	if deleted.DeletedCount > 0 {
-		fmt.Println("Order successfully updated with id: ", id)
+	if deleted.DeletedCount == 0 {
+		return models.Order{}, errors.New("no deleted order")
 	}
 
-	return deletedOrder
+	return deletedOrder, nil
 }
 
-func deleteAllOrders() []models.Order {
+func deleteAllOrders() ([]models.Order, error) {
 
-	deletedOrders := getAllOrders()
+	deletedOrders, _ := getAllOrders()
 
 	deleted, err := orderCollection.DeleteMany(context.Background(), bson.M{})
 	if err != nil {
-		fmt.Println("Unable to delete all orders. Error: ", err)
+		return []models.Order{}, err
 	}
 
-	if deleted.DeletedCount > 0 {
-		fmt.Println("Orders successfully deleted")
+	if deleted.DeletedCount == 0 {
+		return []models.Order{}, errors.New("no deleted record")
 	}
 
-	return deletedOrders
+	return deletedOrders, nil
 }
 
 func GetOneOrder(id string) (models.Order, error) {
 	orderId, _ := primitive.ObjectIDFromHex(id)
-	order := getOneOrder(orderId)
+
+	order, err := getOneOrder(orderId)
+
+	if err != nil {
+		return models.Order{}, err
+	}
 
 	if order == (models.Order{}) {
 		return models.Order{}, errors.New("no such order")
@@ -121,7 +129,11 @@ func GetOneOrder(id string) (models.Order, error) {
 }
 
 func GetAllOrders() ([]models.Order, error) {
-	orders := getAllOrders()
+	orders, err := getAllOrders()
+
+	if err != nil {
+		return []models.Order{}, err
+	}
 
 	if len(orders) > 0 {
 		return []models.Order{}, errors.New("no orders")
@@ -132,21 +144,38 @@ func GetAllOrders() ([]models.Order, error) {
 func CreateOneOrder(order models.Order) error {
 	product, _ := GetOneProduct(order.ProductID.Hex())
 
+	user, _ := GetOneUser(order.UserID.Hex())
+
 	order.Price = product.Price * order.Quantity
 
 	inserted, err := insertOneOrder(order)
 	if err != nil {
 		return err
 	}
-	if inserted != (&mongo.InsertOneResult{}){
-	
+	if inserted != (&mongo.InsertOneResult{}) {
+
 		var updatedValue = primitive.M{
-			"quantity": product.Quantity-order.Quantity,
+			"quantity": product.Quantity - order.Quantity,
 		}
 
-		var updatedProduct = UpdateOneProduct(order.ProductID.Hex(), updatedValue)
-		if updatedProduct != (models.Product{}){
+		updatedProduct, err := UpdateOneProduct(order.ProductID.Hex(), updatedValue)
+		if err != nil {
+			return errors.New("product quantity not updated")
+		}
+		if updatedProduct != (models.Product{}) {
 			return errors.New("product quantity not yet updated")
+		}
+
+		updatedUserValue := primitive.M{
+			"noOfOrders": user.NumberOfOrders + 1,
+		}
+
+		updatedUser, err := UpdateOneUser(order.UserID.Hex(), updatedUserValue)
+		if err != nil {
+			return errors.New("user's no of orders not updated")
+		}
+		if updatedUser != (models.User{}) {
+			return errors.New("user's noOfOrders not yet updated")
 		}
 		return nil
 	} else {
@@ -154,31 +183,50 @@ func CreateOneOrder(order models.Order) error {
 	}
 }
 
-func UpdateOneOrder(id string, updatedValues primitive.M) models.Order {
+func UpdateOneOrder(id string, updatedValues primitive.M) (models.Order, error) {
 	orderId, _ := primitive.ObjectIDFromHex(id)
-	order := getOneOrder(orderId)
+	order, err := getOneOrder(orderId)
 
-	if order != (models.Order{}) {
-		updatedOrder := updateOneOrder(orderId, updatedValues)
-		return updatedOrder
+	if err != nil {
+		return models.Order{}, err
 	}
 
-	return models.Order{}
-}
-
-func DeleteOneOrder(id string) models.Order {
-	orderId, _ := primitive.ObjectIDFromHex(id)
-	order := getOneOrder(orderId)
-
 	if order != (models.Order{}) {
-		deletedOrder := deleteOneOrder(orderId)
-		return deletedOrder
+		updatedOrder, err := updateOneOrder(orderId, updatedValues)
+		if err != nil {
+			return models.Order{}, err
+		}
+		return updatedOrder, nil
 	}
 
-	return models.Order{}
+	return models.Order{}, errors.New("no updated order")
 }
 
-func DeleteAllOrders() []models.Order {
-	deletedOrders := deleteAllOrders()
-	return deletedOrders
+func DeleteOneOrder(id string) (models.Order, error) {
+	orderId, _ := primitive.ObjectIDFromHex(id)
+	order, err := getOneOrder(orderId)
+
+	if err != nil {
+		return models.Order{}, err
+	}
+
+	if order != (models.Order{}) {
+		deletedOrder, err := deleteOneOrder(orderId)
+
+		if err != nil {
+			return models.Order{}, err
+		}
+
+		return deletedOrder, nil
+	}
+
+	return models.Order{}, errors.New("no deleted order")
+}
+
+func DeleteAllOrders() ([]models.Order, error) {
+	deletedOrders, err := deleteAllOrders()
+	if err != nil {
+		return []models.Order{}, err
+	}
+	return deletedOrders, nil
 }
